feat: Add character creation functionality
This commit adds character creation functionality to the game. It includes changes to the `onGameEventTriggered` and `test` controllers to import the `SendCharacterCreate` function from the `createCharacter` module. The `SendCharacterCreate` function is called in the code to send a character create message to the server. Additionally, a new file called `createCharacter.ts` is created in the `nui` module directory, which exports the `SendCharacterCreate` function. This function sets the NUI focus, sends a NUI message to open the register page, and accepts a boolean parameter that determines the state of the creation toggle.

Other changes in this commit include the addition of the `axios` dependency in the `interface` package.json file and changes to the `style.css` file in the `interface` directory for visual adjustments.