feat: Add character creation functionality
This commit adds code for character creation, including the ability to send character creation data from the frontend to the server and insert the newly created character into the user's character data. It also includes a function to count the number of characters for a user and a few new interfaces related to characters and users.